requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pillow==10.1.0
djoser==2.2.0
xhtml2pdf==0.2.11
django-filter==23.3
dateutils==0.6.12
python-decouple==3.8
celery[redis]==5.3.4
flower==2.0.0
celery-singleton==0.3.1
